Beefing up the backend code
This patch adds the mongodb schemas for Users and Coupons entities.

This patch splits the code into the MVC, where model contains the data
representation and the controllers contain the logic for the web
application.

This patch also added users.js to controllers, which actually handles
CRUD REST APIs for the Users entities.